% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3p3.R
\name{m.sim3p3}
\alias{m.sim3p3}
\title{simulates a single dose-escalation studies with modified 3+3 design}
\usage{
m.sim3p3(truerate, targetrate, seed = NULL)
}
\arguments{
\item{truerate}{a vector of true rates for DLT probabilities at different dose levels}

\item{targetrate}{the set target rate of DLT probability to determine the MTD}

\item{seed}{if not empty, the seed to use for random generation}
}
\value{
the output from \code{\link{m.sim3p3}}
}
\description{
simulates a single dose-escalation studies with modified 3+3 design
}
\examples{
m.sim3p3(c(0.25,0.35,0.4,0.45,0.5,0.6), 0.25)
}
