% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3p3.R
\name{m.ssim3p3}
\alias{m.ssim3p3}
\title{simulates n dose-escalation studies with the modified 3+3 design}
\usage{
m.ssim3p3(truerate, targetrate, n, r = 2, seed = NULL)
}
\arguments{
\item{truerate}{a vector of true rates for DLT probabilities at different dose levels}

\item{targetrate}{the set target rate of DLT probability to determine the MTD}

\item{n}{number of studies to simulate}

\item{r}{integer, number of digits for percentages in output}

\item{seed}{if not empty, the seed to use for random generation}
}
\value{
the output from \code{\link{m.ssim3p3}}
}
\description{
simulates n dose-escalation studies with the modified 3+3 design
}
\examples{
m.ssim3p3(c(0.25,0.35,0.4,0.45,0.5,0.6), 0.25, 10000)
}
